requirements.txt
torch==1.9.1
loguru==0.5.3
python-telegram-bot==13.7
made-ai-dungeon==0.0.2
numpy==1.21.2